VerticesExtractor: don't flush while waiting for messages
This patch does not change anything from the plugin's point of
view. However, it adds clarity because otherwise it is not clear
within the corresponding slave mode handler whether or not
flushing should occur in flushAll() if there is any pending message.

+++ code/plugin_VerticesExtractor/VerticesExtractor.cpp
@@ -317,10 +317,18 @@ void VerticesExtractor::setReadiness(){
 	#ifdef ASSERT
 	assert(m_pendingMessages>0);
 	#endif
+
 	m_pendingMessages--;
 }
 
 void VerticesExtractor::flushAll(RingAllocator*m_outboxAllocator,StaticVector*m_outbox,int rank){
+
+/**
+ * wait for the reply for current messages
+ */
+	if(m_pendingMessages)
+		return;
+
 	if(!m_bufferedData.isEmpty()){
 		m_pendingMessages+=m_bufferedData.flushAll(RAY_MPI_TAG_VERTICES_DATA,m_outboxAllocator,m_outbox,m_parameters->getRank());
 		return;
